Perform sse float comparisons with the floating-point intrinsics
This fixes an issue that was causing unit tests to fail on Linux in profile
builds with clang 12, specifically the `MATH_Vector2.TestReciprocal`
test:

```
TEST(MATH_Vector2, TestReciprocal)
{
    AZ_TEST_ASSERT(Vector2(2.0f, 4.0f).GetReciprocal().IsClose(Vector2(0.5f, 0.25f)));
    AZ_TEST_ASSERT(Vector2(2.0f, 4.0f).GetReciprocalEstimate().IsClose(Vector2(0.5f, 0.25f)));
}
```

The second test, `GetReciprocalEstimate()`, is the one that was failing.
`GetReciprocalEstimate()` would return the same value in both clang 6 and
clang 12, but clang 12 would generate different code for `IsClose()`.

`IsClose(Vector2 a, Vector2 b, float tolerance)` works like this:
* subtract b from a, and take the absolute value of the result, store in c
* check that all the elements of c are less than the tolerance

The less than comparison done in `IsClose()` uses simd registers and sse
instructions to vectorize the comparisons. That is done in the function
`Simd::Vec2::CmpAllLtEq`, which on Linux uses the SSE implementation. The
previous implementation was this:

```cpp
AZ_MATH_INLINE bool CmpAllLtEq(__m128 arg1, __m128 arg2, int32_t mask)
{
    const __m128i compare = CastToInt(CmpGt(arg1, arg2));
    return (_mm_movemask_epi8(compare) & mask) == 0;
}
```

This function is used by all 4 vector types. It uses the same instructions
to test a Vector2 as it does a Vector4, and uses the `mask` parameter to
determine which floats from the vector to include in the test.

On clang 6, this function would generate the following instructions:
```asm
cmpltps  xmm0, xmm1
pmovmskb ecx, xmm0
xor      eax, eax
test     cl, cl
```

* `cmpltps` compares the 4 floats in `xmm0` and `xmm1`. If
  `xmm0[0] < xmm1[0]`, it writes `0x00000000` to `xmm0[0]`. Otherwise, it
  writes `0xFFFFFFFF` to `xmm0[0]`. It repeats that for all 4 floats.
* `pmovmskb` creates a mask made up of the most significant bit of each
  byte of the source operand and stores the result in the low byte
  of the destination operand. For our use case, using it on the result of
  `cmpltps`, the result is 4 `1` bits for each element that returned true
  in the cmp, and 4 `0` bits otherwise

So if we have some example:
vector a is (0.50, 0.25, 0.1, 0.1)
vector b is (0.49, 0.25, 0.1, 0.0)
`cmpltps` will produce:
0xffffffff, 0x00000000, 0x00000000, 0xffffffff
Then `pmovmskb` will produce:
0x00f00f

On clang 12, the same function would generate these instructions:
```asm
cmpltps  xmm0, xmm1
movmskps ecx, xmm0
xor      eax, eax
test     ecx, ecx
```

The `cmpltps` is the same, but the mask instruction used is `movmskps`.
This instruction extracts just the sign bits from the floating-point
values in the source and formats them into a 4-bit mask. So taking the same
example:
vector a is (0.50, 0.25, 0.1, 0.1)
vector b is (0.49, 0.25, 0.1, 0.0)
`cmpltps` will produce:
0xffffffff, 0x00000000, 0x00000000, 0xffffffff
Then `movmskps` will produce:
0b00000000000000000000000000001001

I was not able to determine why this behavior changed in clang. However,
this change causes clang and MSVC to use the `movmskps` instruction
instead.

The masks used by each vector type was updated to have only 1 bit per
used element: for Vector1, it is `0b0001`, for Vector 2, it is `0b0011`,
etc.

In addition, the previous version of the code implemented a
less-than-or-equal-to comparison using the greater-than intrinsic, and
checking if the result was false. This new version implements
less-than-or-equal-to using the less-than-or-equal-to intrinsic, and checks
if the result is true.

Signed-off-by: Chris Burel <burelc@amazon.com>