Introduce the O3DEKernel shared library, containing the Environment
This introduces a new top-level library to be the central holder of the
global Environment instance. It removes the need to explicitly Attach and
Detach the Environment, since the Environment will now be available during
static initialization via the exported library functions from this new
shared library.

The Kernel library only contains the Environment code, and the code it
needs from AzCore (namely, `AzCore/std/hash.cpp`).

The decision was made to not include the `AZ::Debug::Trace` library in the
Kernel library, in order to avoid the additional dependencies that `Trace`
pulls in, like `EBus`.

Signed-off-by: Chris Burel <burelc@amazon.com>