Don't export GetModuleId(), so that a unique value per module is used
The first iteration of making the O3DEKernel shared library just exported
all the symbols that are used by the AzCore tests. `GetModuleId()` is one
of them, but by exporting that function, it no longer reuturns a unique
value per module.

This change fixes that behavior, so that a unique value is returned.

The previous implementation used the address of an unrelated variable with
static storage at the global scope, namely `g_environmentCleanUp`. Because
this variable was defined in a static library, it would have a different
address for each module it was linked into. `g_environmentCleanUp` is now
only defined in the shared library, so using it to identify each module is
no longer correct.

Signed-off-by: Chris Burel <burelc@amazon.com>