Change Environment to be constructed on first use
Now that the Environment is provided as a single instance via a shared
library, modules do not need to each maintain a separate reference to the
Environment via a global static pointer. The Environment is guaranteed
to be available to any library that links to O3DEKernel. This means that
`Environment::Attach()` and `Environment::Detach()` don't have anything to
do, and that `IsReady()` can always return `true`. The `Environment`
instance is now created as a function-local variable with static storage,
so that it is initialized the first time it is asked for. The behavior of
statics being destroyed in an atexit handler removes the need for the
`Environment::CleanUp` type.

Signed-off-by: Chris Burel <burelc@amazon.com>