Remove special Environment creation for tests on Android
On Android, `AzTestRunner` is built as a shared library. Previously, in
order to ensure that one single environment instance was used, a function
was exported from that library to get The One Single Environment. Because
that functionality is now the standard way of accessing the Environment,
provided by the O3DEKernel library, this code is no longer needed.

Signed-off-by: Chris Burel <burelc@amazon.com>