Used an optimized double linked list to speed up searching for available indices in AZ::IO::Streamer.
Previously searching for an available slot for caches or file handles required keeping track of the time a slot was last used and search the entire list to find the least recently used. The new approach maintains a double linked list to keep track of the least and most recently used slots. This makes finding a slot much faster, but does require a bit more work when having to touch a file.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>