Fix Json map tests that rely on allocations resulting in a specific address order
AZStd::map is an ordered container. When the key type for a map is a
pointer, then the address that pointer points to is used for the sorting
algorithm for where that item should appear in the map. When a map is
serialized to json, it is merely iterated over in (sorted) order.

The json tests serialize a map out to a string, then compare it against a
known string, so the order of the elements is important. That means that
care must be taken when inserting elements into the map, that things are
inserted in the correct order. For a map of pointers, that means that the
memory address of the allocated objects is what determines the map
element's order.

These tests were previously relying on the allocator to always return a
higher address for later allocations, which is not a guarantee of the
allocator. This fixes that error in the test by allocating the objects
first, inserting them in the map so they are sorted by address, then
traversing the map to set the values appropriately.

Signed-off-by: Chris Burel <burelc@amazon.com>