[ThreadPoolAllocator] Free more pages during GarbageCollect()
This moves the freeing of pages from the beginning of GarbageCollect() to
the end. Calling GarbageCollect on the per-thread PoolAllocator can cause
pages to be added to the free page list, so by freeing pages at the end,
more free pages can be freed in a single GarbageCollect() call.

This fixes leaks reported by the OSAllocator in the AzCore
ThreadPoolAllocatorTest.Test test.

Signed-off-by: Chris Burel <burelc@amazon.com>