Avoid creating an allocator in its `Destroy()` method
It seems some places call `AllocatorInstance<T>::Create()` without ever
using that allocator, and then call `AllocatorInstance<T>::Destroy()`. Now
that `Create()` is a no-op, if there's no intervening call to `Get()`, then
the allocator isn't actually instantiated. Previously, the existing call to
`Destroy()` _would_ call `Get()`, which would create it. That was causing
the allocator to be registered with the manager, which would then cause
tests to issue a failure/warning, stating that there's still a registered
allocator.

This avoids that probelm by simply not creating the allocator in
`Destroy()`. If the allocator is never used, it is not created.

A more complete fix will be to remove the `Create()`, `Destroy()`, and
`IsReady()` methods from `AllocatorInstance`.

Signed-off-by: Chris Burel <burelc@amazon.com>