Add class allocators to subclasses, to ensure the correct sizes are reported on free
If a base class uses the `AZ_CLASS_ALLOCATOR` macro to declare that
instances of a type should be allocated with a specific allocator, then
subclasses of that type also need to use the same macro. Without it, the
size of the subclass is passed to `operator new`, but the size of the
_base_ class is passed to `operator delete`, due to the subclass not having
overridden its `operator delete`. This causes the allocator to have a size
mismatch, resulting in incorrect tracking information.

Signed-off-by: Chris Burel <burelc@amazon.com>